consolidated into single custom variable GUI file
